% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adc_get.R
\name{adc_get}
\alias{adc_get}
\title{Get data from \emph{AgroDataCube}}
\usage{
adc_get(
  url = NULL,
  option,
  params,
  server = "adc",
  download = FALSE,
  token = Sys.getenv("ADC_TOKEN"),
  out_path = tempfile(),
  overwrite = TRUE
)
}
\arguments{
\item{url}{character. Request URL.}

\item{option}{character. Determines the type of request.}

\item{params}{vector or list. List of named parameters.}

\item{server}{character. Server to connect to, either the main or test \emph{AgroDataCube} server.}

\item{download}{boolean. If \code{TRUE}, download file.}

\item{token}{character. API token.}

\item{out_path}{character. Output path.}

\item{overwrite}{boolean. If \code{TRUE}, overwrite file.}
}
\value{
The parsed content of the response (a list for JSON responses). An error is raised if the request fails.
}
\description{
Wrapper to get data from \emph{AgroDataCube} through their REST API.
}
