% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adc_url.R
\name{adc_url}
\alias{adc_url}
\title{Compose \emph{AgroDataCube} URL}
\usage{
adc_url(option, params, base_url = "https://agrodatacube.wur.nl/api/v2/rest/")
}
\arguments{
\item{option}{character. Determines the type of request.}

\item{params}{vector or list. List of named parameters.}

\item{base_url}{character. Base URL for the REST API.}
}
\value{
A request URL text string.
}
\description{
Compose URL text string to submit requests through \emph{AgroDataCube} REST APIs.
}
