% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndc_trange.R
\name{as_rfc3339}
\alias{as_rfc3339}
\title{Date(s) in RFC 3339 format}
\usage{
as_rfc3339(x)
}
\arguments{
\item{x}{Date, POSIXct or character. One or more dates.}
}
\value{
character. The dates in RFC 3339 format.
}
\description{
Convert dates to the RFC 3339 format (UTC).
}
