% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets_download_wcs.R
\name{assets_download_wcs}
\alias{assets_download_wcs}
\title{Download STAC assets originating in WCS servers}
\usage{
assets_download_wcs(
  items,
  asset_names = "wcs",
  output_dir = tempdir(),
  output_ext = ".tif",
  overwrite = TRUE
)
}
\arguments{
\item{items}{doc_items. Items list resulting from \code{ndc_get}.}

\item{asset_names}{character. Names of the STAC assets to download.}

\item{output_dir}{character. Output directory path.}

\item{output_ext}{character. Output file extension.}

\item{overwrite}{boolean. If \code{TRUE}, overwrite file.}
}
\value{
The \code{httr} response (or list of responses), or \code{NULL} if there is nothing to download.
}
\description{
Wrapper to download data from STAC endpoints that originated in WCS servers.
}
