% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_url.R
\name{gm_url}
\alias{gm_url}
\title{Compose \emph{GroenMonitor} URL}
\usage{
gm_url(option = "NDVI", params)
}
\arguments{
\item{option}{character. Determines the type of data.}

\item{params}{vector or list. List of named parameters.}
}
\value{
A request URL text string.
}
\description{
Compose URL text string to submit requests to the \emph{GroenMonitor} WCS server.
}
