% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndc_count.R
\name{ndc_count}
\alias{ndc_count}
\title{Count datasets within spatiotemporal contraints}
\usage{
ndc_count(
  s = NULL,
  collection = NULL,
  roi = NULL,
  trange = NULL,
  limit = 100,
  token = Sys.getenv("NDC_TOKEN")
)
}
\arguments{
\item{s}{doc_items. Search result (e.g. from \code{\link[=ndc_get]{ndc_get()}}). If missing, a new search is made with the other arguments.}

\item{collection}{character. Collection ID.}

\item{roi}{character, numeric or sf. Region of interest, see \code{\link[=ndc_roi]{ndc_roi()}}. Can be either: (i) a path to a file with a custom geometry, (ii) a numeric vector with coordinates representing a bounding box, or (iii) an sf object with a (multi)polygon representing a custom region of interest.}

\item{trange}{character. Temporal range.}

\item{limit}{integer. Maximum number of STAC Items to return.}

\item{token}{character. API token.}
}
\value{
integer. Number of STAC Items matched by the search query.
}
\description{
Obtain a number of items available in NatureDataCube datasets (optionally constrained by query parameters).
}
