% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndc_datasets.R
\name{ndc_datasets}
\alias{ndc_datasets}
\title{List available datasets}
\usage{
ndc_datasets(
  roi = NULL,
  trange = NULL,
  token = Sys.getenv("NDC_TOKEN"),
  total = FALSE,
  matched = FALSE
)
}
\arguments{
\item{roi}{character, numeric or sf. Region of interest, see \code{\link[=ndc_roi]{ndc_roi()}}. Can be either: (i) a path to a file with a custom geometry, (ii) a numeric vector with coordinates representing a bounding box, or (iii) an sf object with a (multi)polygon representing a custom region of interest.}

\item{trange}{character. Temporal range.}

\item{token}{character. API token.}

\item{total}{boolean. If \code{TRUE}, return total number of STAC Items for each STAC Collection.}

\item{matched}{boolean. If \code{TRUE}, return matched number of STAC Items for each STAC Collection.}
}
\value{
character or data frame. List of STAC Collections and (optionally) the total and/or matched STAC Items count.
}
\description{
List all datasets (optionally constrained by query parameters) in NatureDataCube.
}
