% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndc_sites.R
\name{ndc_sites}
\alias{ndc_sites}
\title{Load study sites}
\usage{
ndc_sites(layer = NULL)
}
\arguments{
\item{layer}{character. Pre-defined study site. If empty (or \code{layer = NULL}), return all available study sites. Otherwise, return the boundaries for the pre-defined study site with that name.}
}
\value{
Either a character vector with layer names, or an sf object with a (multi)polygon representing the boundaries for a pre-defined study site.
}
\description{
Load polygons for all study sites available through LTER-LIFE.
}
