% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rNDC-package.R
\docType{package}
\name{rNDC-package}
\alias{rNDC}
\alias{rNDC-package}
\title{rNDC: R interface to LTER-LIFE's NatureDataCube}
\description{
Functions and wrappers to find and get data from the NatureDataCube STAC API,
and from the related AgroDataCube and GroenMonitor services.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LTER-LIFE/rNDC}
  \item Report bugs at \url{https://github.com/LTER-LIFE/rNDC/issues}
}

}
\author{
\strong{Maintainer}: Bruno Marcos \email{b.a.santosmarcos@uva.nl}

Authors:
\itemize{
  \item Daan Lichtenberg \email{D.Lichtenberg@nioo.knaw.nl}
  \item Minke Mulder
}

}
\keyword{internal}
