% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_date_range.R
\name{split_date_range}
\alias{split_date_range}
\title{Split a date range}
\usage{
split_date_range(fromdate, todate, by_days = 7)
}
\arguments{
\item{fromdate, todate}{Date or character. Start and end of the range.}

\item{by_days}{integer. Length (in days) of each chunk.}
}
\value{
A data frame with the \code{from} and \code{to} dates of each chunk.
}
\description{
Split a date range into consecutive chunks.
}
