% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndc_roi.R
\name{ndc_roi}
\alias{ndc_roi}
\title{Spatial region of interest}
\usage{
ndc_roi(roi = NULL)
}
\arguments{
\item{roi}{character, numeric or sf. Region of interest. Can be either: (i) a character pointing to a path of a file containing a custom geometry, (ii) a numeric vector with four coordinates representing a bounding box, or (iii) an sf object with a (multi)polygon representing a custom region of interest.}
}
\value{
An sfc object in EPSG:4326 (or \code{NULL} if \code{roi} is \code{NULL}). If the RoI has no CRS, EPSG:4326 is assumed.
}
\description{
Import and transform spatial region of interest.
}
